requirements.txt
azure-identity
azure-monitor-query
pyinstaller